+++ src/AppLazy.tsx
@@ -3,14 +3,16 @@ import { ReactFlowProvider } from "@xyflow/react";
 import "@xyflow/react/dist/style.css";
 import { Toaster } from "react-hot-toast";
 import { App } from "./App";
+import { OverlayProvider } from "./core/ui/overlayState";
 
 export default function AppLazy() {
   return (
     <ReactFlowProvider>
-      <App />
-      <Toaster />
+      <OverlayProvider>
+        <App />
+        <Toaster />
+      </OverlayProvider>
     </ReactFlowProvider>
   );
 }
-